feat: add photo conversion utility to Base64 and integrate it into user profile serialization
