+++ api_usuario/serializers.py
@@ -1,19 +1,7 @@
 from rest_framework import serializers
 from .models import Usuario, Credencial
 from datetime import date
-import base64
-
-# FUNÇÃO AUXILIAR PARA CONVERTER BINÁRIO (BYTEA) PARA BASE64
-def converter_foto_para_base64(instancia_foto):
-    if instancia_foto:
-        try:
-            # Converte o memoryview do Postgres para bytes e depois para string Base64
-            foto_bytes = bytes(instancia_foto)
-            foto_base64 = base64.b64encode(foto_bytes).decode('utf-8')
-            return f"data:image/jpeg;base64,{foto_base64}"
-        except Exception:
-            return None
-    return None
+from utils.converter_foto import converter_foto_para_base64
 
 # SERIALIZERS PARA O CADASTRO DE USUÁRIOS COM VALIDAÇÕES
 class CadastroSerializer(serializers.Serializer):
@@ -68,7 +56,6 @@ class PerfilSerializer(serializers.ModelSerializer):
 
     def get_stats(self, obj):
         return {
-            "postagens": 42,
             "seguindo": 1250,
             "seguidores": 108
         }
